# packages/api-server/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Rita API",
		"version": "1.0.0",
		"description": "Rita API Server for the RITA application.\n\n## Authentication\nMost endpoints require authentication via Keycloak JWT token or session cookie.\n\n## Organization Context\nAll data is scoped to the user's active organization (tenant).\n\n## Real-time Updates\nFor real-time updates, connect to the SSE endpoint at `/api/sse/events`.\nSee AsyncAPI spec for event documentation.",
		"contact": {
			"name": "Rita Team"
		}
	},
	"servers": [
		{
			"url": "/",
			"description": "Current server"
		}
	],
...
	"security": [
		{
			"bearerAuth": []
		},
		{
			"cookieAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Keycloak JWT token"
			},
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session",
				"description": "Session cookie authentication"
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"description": "Error message",
						"example": "Resource not found"
					},
					"code": {
						"type": "string",
						"description": "Error code",
						"example": "NOT_FOUND"
					},
					"details": {
						"description": "Additional error details"
					}
				},
				"required": ["error"]
			},
...
			"DataSourceVerifyResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string",
						"example": "Verification initiated"
					},
					"connection_id": {
						"type": "string",
						"format": "uuid"
					}
				},
				"required": ["success", "message", "connection_id"]
			},
			"DataSourceSyncResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string",
						"example": "Sync initiated"
					},
					"connection_id": {
						"type": "string",
						"format": "uuid"
					}
				},
				"required": ["success", "message", "connection_id"]
			},
			"DataSourceSyncTicketsResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string",
						"example": "Ticket sync initiated"
					},
					"connection_id": {
						"type": "string",
						"format": "uuid"
					},
					"ingestion_run_id": {
						"type": "string",
						"format": "uuid"
					}
				},
				"required": ["success", "message", "connection_id", "ingestion_run_id"]
			},
...
		},
...
	},
	"paths": {
...
		"/api/data-sources/{id}/verify": {
			"post": {
				"tags": ["Data Sources"],
				"summary": "Verify credentials",
				"description": "Initiate credential verification for a data source. Returns immediately with status \"verifying\". Final result arrives via SSE event `data_source_update`.",
				"security": [
					{
						"bearerAuth": []
					},
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"schema": {
							"type": "string",
							"format": "uuid",
							"description": "Data source connection ID"
						},
						"required": true,
						"description": "Data source connection ID",
						"name": "id",
						"in": "path"
					}
				],
				"responses": {
					"200": {
						"description": "Verification initiated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataSourceVerifyResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Data source not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/data-sources/{id}/sync": {
			"post": {
				"tags": ["Data Sources"],
				"summary": "Trigger sync",
				"description": "Initiate document sync from the data source. Returns immediately with status \"syncing\". Progress and completion arrive via SSE event `data_source_update`.",
				"security": [
					{
						"bearerAuth": []
					},
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"schema": {
							"type": "string",
							"format": "uuid",
							"description": "Data source connection ID"
						},
						"required": true,
						"description": "Data source connection ID",
						"name": "id",
						"in": "path"
					}
				],
				"responses": {
					"200": {
						"description": "Sync initiated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataSourceSyncResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Data source not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/data-sources/{id}/sync-tickets": {
			"post": {
				"tags": ["Data Sources"],
				"summary": "Sync ITSM tickets",
				"description": "Initiate ITSM ticket ingestion for Autopilot clustering. Only valid for ITSM data sources (servicenow_itsm, jira_itsm, freshservice_itsm, ivanti_itsm). Returns HTTP 202 Accepted. Progress arrives via SSE event `ingestion_run_update`.",
				"security": [
					{
						"bearerAuth": []
					},
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"schema": {
							"type": "string",
							"format": "uuid",
							"description": "ITSM data source connection ID"
						},
						"required": true,
						"description": "ITSM data source connection ID",
						"name": "id",
						"in": "path"
					}
				],
				"responses": {
					"202": {
						"description": "Ticket sync accepted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataSourceSyncTicketsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid data source type (not ITSM)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Data source not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}